+++ src/main/java/com/example/pillyohae/order/entity/Order.java
@@ -3,7 +3,7 @@ package com.example.pillyohae.order.entity;
 import com.example.pillyohae.coupon.entity.CouponTemplate;
 import com.example.pillyohae.coupon.entity.IssuedCoupon;
 import com.example.pillyohae.global.entity.BaseTimeEntity;
-import com.example.pillyohae.global.entity.adress.ShippingAddress;
+import com.example.pillyohae.global.entity.address.ShippingAddress;
 import com.example.pillyohae.order.entity.status.OrderItemStatus;
 import com.example.pillyohae.order.entity.status.OrderStatus;
 import com.example.pillyohae.user.entity.User;
